+++ app/src/main/java/com/example/software_engineering_project/adapter/CreateFlatShareViewAdapter.java
@@ -1,4 +1,4 @@
-package com.example.software_engineering_project;
+package com.example.software_engineering_project.adapter;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -11,6 +11,9 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.example.software_engineering_project.controller.CreateFlatShareScreen;
+import com.example.software_engineering_project.R;
 
 import java.util.ArrayList;
