+++ app/src/main/java/com/example/software_engineering_project/controller/RegisterScreen.java
@@ -1,10 +1,12 @@
-package com.example.software_engineering_project;
+package com.example.software_engineering_project.controller;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Button;
+
+import com.example.software_engineering_project.R;
 
 public class RegisterScreen extends AppCompatActivity {
