+++ app/src/main/res/layout/activity_change_mail_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".ChangeMailScreen">
+    tools:context=".controller.ChangeMailScreen">
 
     <RelativeLayout
         android:layout_width="match_parent"
