+++ app/src/main/res/layout/activity_change_password_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".ChangePasswordScreen">
+    tools:context=".controller.ChangePasswordScreen">
 
     <RelativeLayout
         android:layout_width="match_parent"
