+++ app/src/main/res/layout/activity_grocery_list_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GroceryListMain">
+    tools:context=".controller.GroceryListMain">
 
     <RelativeLayout
         android:layout_width="match_parent"
