+++ app/src/main/res/layout/activity_login_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".LoginScreen">
+    tools:context=".controller.LoginScreen">
 
     <RelativeLayout
         android:layout_width="match_parent"
