+++ app/src/main/res/layout/activity_main_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".MainScreen">
+    tools:context=".controller.MainScreen">
 
     <RelativeLayout
         android:layout_width="match_parent"
