+++ app/src/main/res/layout/activity_manage_flat_share_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".ManageFlatShareMain">
+    tools:context=".controller.ManageFlatShareMain">
 
     <RelativeLayout
         android:layout_width="match_parent"
