+++ app/src/main/res/layout/activity_setting_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".SettingScreen">
+    tools:context=".controller.SettingScreen">
 
     <RelativeLayout
         android:layout_width="match_parent"
