+++ app/src/main/res/values-de/strings.xml
@@ -28,6 +28,6 @@
     <string name="enter_item_here">Artikel hier eintragen</string>
     <string name="name_of_flat">Name der WG</string>
     <string name="enter_mail_here">Mail-Adresse hier eintragen</string>
-    <string name="save_changes">Änderungen speichern</string>
+    <string name="save_changes">Speichern</string>
     <string name="enter_flat_share_name">Name der WG hier eintragen</string>
 </resources>
